fix the value type

+++ lab1.c
@@ -10,7 +10,7 @@ int main() {
   char *word = "";
   if (read != -1) {
 
-    for (int i = 0; i < strlen(sentence); i++) {
+    for (size_t i = 0; i < strlen(sentence); i++) {
       if (sentence[i] == ' ') {
         printf("%s \n", word);
         word = "";
